[feat] #5-db 엔티티 추가 생성
