+++ src/main/java/servnow/servnow/domain/answer/repository/MultipleChoiceRepository.java
@@ -1,0 +1,7 @@
+package servnow.servnow.domain.answer.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MultipleChoiceRepository extends JpaRepository<MultipleChoice, Long> {
+
+}
