+++ src/main/java/servnow/servnow/domain/answerResult/model/AnswerResult.java
@@ -1,0 +1,24 @@
+package servnow.servnow.domain.answerResult.model;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import servnow.servnow.domain.common.BaseTimeEntity;
+import servnow.servnow.domain.surveyResult.model.SurveyResult;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class AnswerResult extends BaseTimeEntity{
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long answerId;
+
+    @ManyToOne
+    @JoinColumn(name = "answer_result_id", nullable = false)
+    private SurveyResult surveyResult;
+
+    private Long subjectiveId;
+    private Long multipleChoiceId;
+}
