+++ src/main/java/servnow/servnow/domain/answerResult/repository/AnswerResultRepository.java
@@ -1,0 +1,8 @@
+package servnow.servnow.domain.answerResult.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import servnow.servnow.domain.answerResult.model.AnswerResult;
+
+public interface AnswerResultRepository extends JpaRepository<AnswerResult, Long> {
+
+}
