+++ src/main/java/servnow/servnow/domain/section/repository/SectionRepository.java
@@ -1,0 +1,8 @@
+package servnow.servnow.domain.section.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import servnow.servnow.domain.section.model.Section;
+
+public interface SectionRepository extends JpaRepository<Section, Long> {
+
+}
